Bug 488315 - [Codegen C++] Public members not under "public" section when there is an inner classifier of public visibility
We set the current visibility to null before going into an inner class
so it can be generated correctly. After processing the inner class, we
reset the current visibility to the previous visibility before going
into an inner classifier. Indeed, when we go into the inner classifier,
the visibility will probably have changed, which results in incoherence.

Change-Id: I99ed27a78dba3560fdb15bfc1156e73c4eb14fe0
Signed-off-by: Shuai Li <shuai.li@cea.fr>